test-language: fix unit test
Compare style_ids strv's after being sorted.

By reading the GtkSourceLanguage API, some strv's are guaranteed to have
the same content order as defined in the *.lang file. But for the
style_ids it's not the case, and the unit test broke after a change in
GLib (an implementation detail in the GLib hash table, apparently). The
unit test was wrongly implemented anyway.