data/style-schemes: improve the scripts and automatic test
Note that "$@" is not the same as $@, check your shell manual if you are
not familiar with it. "$@" is treated specially by the shell.
(it supports whitespace in file names).

The styles.rng file is no longer installed, just distributed in
tarballs. Because the scripts are anyway present only in git or
tarballs, they were and are not installed.

test.sh can be called from another dir, not necessarily from the
directory it is located in. Necessary for `make distcheck`.