StyleScheme: bug fix, simplify get/set_parent*()
In set_parent(), there was a bug to assign the new GObject, in case we
set the same GObject as already stored in scheme->priv->parent (the
object could be freed after g_clear_object()).

g_set_object() prevents this kind of issue.

Also, remove some of the docs, as it's already documented in docs/html/
for the file format (it's already clear to know what is a parent
scheme).