unit-tests: styleschememanager: test unknown-parent differently
To see the code coverage.

`make check-code-coverage` didn't show the new coverage with the
g_warning() being executed.

Now it's the case. And we can check that the schemes GList has length 0
(so the scheme has been removed).

But g_test_assert_expected_messages() is more or less deprecated, since
it doesn't work with G_LOG_USE_STRUCTURED as explained in its docs.

Ideally we want to receive a GError, but it requires to change the
GtkSourceStyleSchemeManager API (or at leats having an internal API for
that).