Style: apply(): read infos from GtkSourceStyleData
Remove the fixme comment because it's a bit hard to fix, it's difficult
to know how to do otherwise.

Remove also the const for the first parameter. Most GObject classes
don't care about const / non-const for the self parameter.

By depending on GtkSourceStyleData, the apply() function is already
migrated. It's planned for the GtkSourceStyle struct to contain only a
GtkSourceStyleData for the data.