StyleSchemeParser: implement new parser based on GMarkupParser
The previous implementation used the libxml2. Maybe one day this library
will no longer depend on the libxml2 (to have fewer dependencies and a
smaller package size on certain platforms).

GMarkupParser is also more familiar as an API for a project like this
one which heavily uses GLib.

The new parser implemented here is a bit more strict than the previous
one. It checks more things.