language-specs: remove testv1.lang
testv1.lang was a test for the old *.lang format.

When trying to use it, it prints the following warning:

```
The language definition format version 1 (for syntax highlighting) is no
longer supported by GtkSourceView >= 4.0. Please use version 2.
```

With Autotools, testv1.lang was not installed, it was just there in the
directory and ignored.