style-schemes: remove support for line-background <style> attribute
It was used in only one very precise place:
kate.xml for the t2t.lang.

If t2t was used with another style scheme, the feature wasn't used.

The feature was not mentioned in the file format docs.

To bring this feature back, it would be more useful to have some general
means of supporting any GtkTextTag property, without duplicating the
data fields in GtkSourceStyle. That would be more intelligent.

In the meantime, having a simpler implementation is useful (less chances
to have bugs).