style-schemes: remove version attribute: docs
The format has evolved (e.g. with the kind attribute), so the version
string would need to be updated.

But the format could be improved further, for example to better support
translatable strings.

It's actually simpler to have no version attribute at all. The parser
gives error messages if the file cannot be read.